chore(web): rename tabs pages into index for shorter import
